refactor(dftu): move cal_occ_mat_k/gamma declarations into dftu_lcao_occ.h
Rename dftu_occup.h/.cpp to dftu_lcao_occ.h/.cpp so the occupation-matrix
declarations (template cal_occ_mat<T> dispatcher plus the two concrete
free functions) live in one header, and drop the now-unneeded hamilt
forward declaration from dftu_lcao.h. Build files updated accordingly.

Verified: make -j 30 in build_max_para_test (exit 0, abacus_max_para
v3.11.0-beta8); agent_governance_check.py --staged passed with warnings
only (header includes required by moved declarations).
