+++ source/source_lcao/module_dftu/CMakeLists.txt
@@ -4,7 +4,7 @@ list(APPEND objects
       dftu_yukawa.cpp
       dftu_folding.cpp
       dftu_tools.cpp
-      dftu_occup.cpp
+      dftu_lcao_occ.cpp
       dftu_hamilt.cpp
 )
